feat(packages/sui-bundler): rename module rules
